Perubahan ruoting dari Product Controller ke StockController

+++ routes/web.php
@@ -20,35 +20,32 @@ Route::group(['middleware' => ['auth']], function () {
 
     // Home Routes 
     Route::get('/', 'HomeController@index');
-    Route::get('/stock', 'HomeController@stock');
     Route::get('/history', 'HomeController@history');
     Route::get('/history/{$id}', 'HomeController@historyShow');
     // End Home
 
-    // Product Routes
-    Route::get('/product', 'ProductController@index');
-    Route::get('/product/branch/{order}', 'ProductController@index');
-    Route::get('/product/excel', 'ProductController@Excel');
-    Route::get('/product/create', 'ProductController@create');
-    Route::get('/product/addStock', 'ProductController@addStock');
-    Route::post('/product/store', 'ProductController@store');
+    // Stock Routes
+    Route::get('/stock', 'StockController@index');
+    Route::get('/stock/branch/{code}', 'StockController@stock');
+    Route::get('/stock/excel/{code}', 'StockController@Excel');
+    Route::get('/stock/create/{code}', 'StockController@create');
+    Route::post('/stock/store/{code}', 'StockController@store');
 
-    Route::get('/product/{product:slug}', 'ProductController@show');
-    Route::delete('/product/delete', 'ProductController@destroy');
-    Route::get('/product/{product:id}/edit', 'ProductController@edit');
-    Route::patch('/product/{product:id}/update', 'ProductController@update');
-    Route::get('/product/{product:id}/market', 'ProductController@market');
-    Route::patch('/product/{product:id}/marketex', 'ProductController@marketex');
-    // End Product
-
+    Route::get('/stock/{product:slug}', 'StockController@show');
+    Route::delete('/stock/delete', 'StockController@destroy');
+    Route::get('/stock/{product:id}/edit', 'StockController@edit');
+    Route::patch('/stock/{product:id}/update', 'StockController@update');
+    Route::get('/stock/{product:id}/market', 'StockController@market');
+    Route::patch('/stock/{product:id}/marketex', 'StockController@marketex');
+    // End Stock
 
     // Branch Routes
-    Route::get('/branch', 'BranchController@index');
-    Route::get('/branch/create', 'BranchController@create');
-    Route::post('/branch/store', 'BranchController@store');
-    Route::get('/branch/{branch:code}', 'BranchController@show');
-    Route::delete('/branch/{branch}/delete', 'BranchController@destroy');
-    Route::get('/branch/{branch:code}/edit', 'BranchController@edit');
-    Route::patch('/branch/{branch:code}/edit', 'BranchController@update');
+    Route::get('/branch', 'BranchController@index')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::get('/branch/create', 'BranchController@create')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::post('/branch/store', 'BranchController@store')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::get('/branch/{branch:code}', 'BranchController@show')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::delete('/branch/{branch}/delete', 'BranchController@destroy')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::get('/branch/{branch:code}/edit', 'BranchController@edit')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
+    Route::patch('/branch/{branch:code}/edit', 'BranchController@update')->middleware('check_user_role:' . \App\Role\UserRole::ROLE_MASTER);
     // End Branch
 });
